src/requirements.txt
Flask==2.3.3
flask-swagger-ui==4.11.1
numpy==1.26.0
pandas==2.1.0
PyYAML==6.0.1

